# realworld-backend/src/articles/articles.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  Param,
  ParseIntPipe,
  Patch,
  Post,
  UseGuards,
} from '@nestjs/common';
import { GetUser } from '../auth/decorator';
import { JwtGuard } from '../auth/guard';
import { ArticleService } from './article.service';
import {
  CreateArticleDto,
  EditArticleDto,
} from './dto';

@UseGuards(JwtGuard)
@Controller('articles')
export class ArticleController {
  constructor(
    private articleService: ArticleService,
  ) { }

  @Get()
  getArticles(@GetUser('id') userId: number) {
    return this.articleService.getArticles(
      userId,
    );
  }

  @Get(':id')
  getArticleById(
    @GetUser('id') userId: number,
    @Param('id', ParseIntPipe) articleId: number,
  ) {
    return this.articleService.getArticleById(
      userId,
      articleId,
    );
  }

  @Post()
  createArticle(
    @GetUser('id') userId: number,
    @Body() dto: CreateArticleDto,
  ) {
    return this.articleService.createArticle(
      userId,
      dto,
    );
  }

  @Patch(':id')
  editArticleById(
    @GetUser('id') userId: number,
    @Param('id', ParseIntPipe) articleId: number,
    @Body() dto: EditArticleDto,
  ) {
    return this.articleService.editArticleById(
      userId,
      articleId,
      dto,
    );
  }

  @HttpCode(HttpStatus.NO_CONTENT)
  @Delete(':id')
  deleteArticleById(
    @GetUser('id') userId: number,
    @Param('id', ParseIntPipe) articleId: number,
  ) {
    return this.articleService.deleteArticleById(
      userId,
      articleId,
    );
  }
}